+++ database/migrations/2026_05_21_095446_create_reports_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -14,8 +13,8 @@ return new class extends Migration
         Schema::create('reports', function (Blueprint $table) {
             $table->id();
             $table->text('report');
-            $table->integer('user_id');
-            $table->integer('reported_user_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('reported_user_id');
             $table->enum('type', ['spam', 'harassment', 'inappropriate', 'copyright', 'misinformation']);
             $table->enum('priority', ['low', 'medium', 'high']);
             $table->enum('status', ['pending', 'in_review', 'resolved', 'dismissed']);
